move iri creation functions to HelperFunctions.ts
