--- www/src/types/OperationStatus.ts
@@ -1,7 +1,0 @@
-enum OperationStatus {
-  SUCCESS,
-  ALREADY_IN_OTHER_PACKAGE,
-  OVERLAP_WITH_OTHER_PACKAGE,
-}
-
-export default OperationStatus;
