+++ www/src/utils/HelperFunctions.ts
@@ -1,5 +1,6 @@
 import { CommissioningPackageContextProps } from "../context/CommissioningPackageContext.tsx";
 import { assetIri } from "./Triplestore.ts";
+import { PipingNetworkSegmentProps } from "../types/diagram/Piping.ts";
 
 export function ensureArray<T>(value: T | T[]): T[] {
   return Array.isArray(value) ? value : [value];
@@ -20,3 +21,25 @@ export const isInPackage = (
   context.activePackage.nodeIds
     ? context.activePackage.nodeIds.includes(assetIri(id))
     : false;
+
+export function iriFromSvgNode(id: string) {
+  return `https://assetid.equinor.com/plantx#${id}`;
+}
+
+export function iriFromPiping(segment: PipingNetworkSegmentProps) {
+  if (
+    segment.Connection?.ToID &&
+    (!segment.PipingComponent || !segment.PropertyBreak)
+  ) {
+    return `https://assetid.equinor.com/plantx#${segment.Connection.ToID}-node${segment.Connection.ToNode}-connector`;
+  } else if (segment.PipingComponent[1]) {
+    return `https://assetid.equinor.com/plantx#${segment.PipingComponent[1].ID}-node2-connector`;
+  } else if (segment.Connection?.FromID) {
+    return `https://assetid.equinor.com/plantx#${segment.Connection.FromID}-node${segment.Connection.FromNode}-connector`;
+  } else if (segment.Connection?.ToID) {
+    return `https://assetid.equinor.com/plantx#${segment.Connection!.ToID}-node${segment.Connection!.ToNode}-connector`;
+  } else {
+    console.error("Something went wrong with iri creation");
+    return ``;
+  }
+}
